minor fixes 2
Путь от начала до конца строится правильно и заносится в вектор path
Но перемещение самого объекта Unit сделано неправильно

+++ background.h
@@ -31,7 +31,7 @@ public:
 
     void addUnit();
 
-    QGraphicsItem* getItem(int x, int y);
+    QGraphicsItem* getItem(size_t x, size_t y);
    // Unit* getUnit(int x, int y);
 
     std::vector<std::vector<int> > map;
